% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too_many_plots_function.R
\name{too_many_plots}
\alias{too_many_plots}
\title{multiplotter}
\usage{
too_many_plots(input_data, cutoff, dataset_name)
}
\arguments{
\item{input_data}{a dataframe containing any number of variables}

\item{cutoff}{a positive number betwen 0 and 1 which is the lower limit of poitive or negative coefficents}

\item{dataset_name}{The name of the data set used.  This is important for keeping track of everything}
}
\description{
This function is used to produce a quick series of charts based on correlation coefficents.  It
can be used to get a quick overveiw of a dataset and what variables display some co-variation.  It is a 
work in progress, and will eventually be able to do more advanced things than it can at the moment.
}
\details{
The initial idea for this project was to create a function which would simply print
a whole bunch of x,y plots at once.  However, upon further investigation this won't be that 
useful when you have more than 10 variables (which would print 45 individual plots at once).  
Consequently I decided that it would be nessicary to narrow down the results a bit so things don't get
out of hand.  I decided the best way would be to calculate the pearson coeffiecints 
and use that to figure out the best plots to use.
}
\examples{
too_many_plots()
}
\keyword{stats}
